1. Для monorepo добавлена проверка на существования модуля в директории с модулями. Так как могут быть установлены внешние зависимости, которые будут в всписке модулей, а директории с ними отсутствуют. 2. Добавлена проверка на наличие установленных зависимостей из списка декларированных. Так как часто бывает, что разрабатывается модуль, и зависимости деклалированны в sync_project_deps и все хорошо работает, так как зависимые модули слинкованны в глобальный node_modules. Но если установить этот модуль на другую машину, где нет слинкованных модулей, то модуль работать не будет. И по этому необходимо проверять, имеются ли модули в devDeps, deps, peerDeps. 3. Добавлена процедура для принудительной проверки версии модулей в npm registry. При запуске процедуры, будут обновлены версии имеющихся модулей. А если модуля нет в списке зависимостей то он будет установлен если декларирован в sync_project_deps.

+++ src/components/mono_repository/install.ts
@@ -3,6 +3,7 @@ import * as spawn from "cross-spawn";
 import * as path from "path";
 import { envConfig } from "../../config/env";
 import { RearguardConfig } from "../../config/rearguard/RearguardConfig";
+import { install_declared_deps } from "../project_deps/install_declared_deps";
 
 export async function install(CWD: string) {
   const { pkg } = new RearguardConfig(envConfig, path.resolve(CWD, "package.json"));
@@ -13,34 +14,38 @@ export async function install(CWD: string) {
 
   process.env.NODE_ENV = "development";
 
-  const result = spawn.sync("npm", ["install"], {
-    cwd: CWD,
-    encoding: "utf8",
-    stdio: "inherit",
-  });
+  const wasInstalled: boolean = await install_declared_deps(CWD);
 
-  if (result.signal) {
-    if (result.signal === "SIGKILL") {
-      console.log(
-        chalk.red(
-          "The build failed because the process exited too early. " +
-            "This probably means the system ran out of memory or someone called `kill -9` on the process.",
-        ),
-      );
+  if (!wasInstalled) {
+    const result = spawn.sync("npm", ["install"], {
+      cwd: CWD,
+      encoding: "utf8",
+      stdio: "inherit",
+    });
 
-      process.exit(1);
-    } else if (result.signal === "SIGTERM") {
-      console.log(
-        chalk.bold.red(
-          "The build failed because the process exited too early. " +
-            "Someone might have called `kill` or `killall`, or the system could be shutting down.",
-        ),
-      );
+    if (result.signal) {
+      if (result.signal === "SIGKILL") {
+        console.log(
+          chalk.red(
+            "The build failed because the process exited too early. " +
+              "This probably means the system ran out of memory or someone called `kill -9` on the process.",
+          ),
+        );
 
-      process.exit(1);
+        process.exit(1);
+      } else if (result.signal === "SIGTERM") {
+        console.log(
+          chalk.bold.red(
+            "The build failed because the process exited too early. " +
+              "Someone might have called `kill` or `killall`, or the system could be shutting down.",
+          ),
+        );
+
+        process.exit(1);
+      }
+
+      process.exit(0);
     }
-
-    process.exit(0);
   }
 
   process.env.NODE_ENV = NODE_ENV;
